% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASEM_newcoin}
\alias{ASEM_newcoin}
\title{ASEM coin}
\format{
A "coin" class object
}
\source{
\url{https://bluefoxr.github.io/COINr/reference/build_example_coin.html}
}
\usage{
ASEM_newcoin
}
\description{
This is an example "coin" class object from the \emph{newer} COINr >v1.0. It is to be
used for examples with the \code{\link[=coin_to_COIN]{coin_to_COIN()}} function.
}
\keyword{datasets}
